v.0.9.13 Statistics: fix and refactoring. Database content housekeeping
StatisticsPage: refactoring and fixes, using the fixed methods in GamonStatistics
  GamonStatistics: fixed the methods for calculations of Means in time bands
  refactored with binding and ViewModel, code behind with no code
  BL_BolusesAndInjections: eliminated bugged code performing statistics. Fixed functions are now in GamonStatistics
  SqLite_BolusesAndInjections: fixed the query that retrieves the rapic action insulin injections
  WORK IN IN PROGRESS FOR "TOTAL" STATISTICS
Outliers e and database error fixes: added functions for finding errori in the database. Ancora nessun outlier statistico
  StatisticsAndGraphPage: added button that help to cure the contant of the database
  btnCureChoMismatchInMeals: check the Meals records to find and possibly correct those that have a difference between the stored CHO of the meal and the sum of foods' eaten CHOs
RecipesPage: fix of a Unit of Measurement not working properly
Hours of the meals: Used in the statistics' operation the hours that are shown in settings

+++ SharedModel/StatisticsData.cs
@@ -9,6 +9,6 @@ namespace GlucoMan
         internal double? DailyMean { get; set; }
         internal double? Mean { get; set; }
         internal double? StandardDeviation { get; set; }
-        internal double? NSamples { get; set; }
+        internal int? NSamples { get; set; }
     }
 }
